sample buffer
