+++ python/cugraph-dgl/cugraph_dgl/dataloading/neighbor_sampler.py
@@ -194,9 +194,9 @@ class NeighborSampler(Sampler):
 
         if g.is_homogeneous:
             indices = torch.concat(list(indices))
-            ds.sample_from_nodes(indices, batch_size=batch_size)
+            reader = ds.sample_from_nodes(indices, batch_size=batch_size)
             return HomogeneousSampleReader(
-                ds.get_reader(), self.output_format, self.edge_dir
+                reader, self.output_format, self.edge_dir
             )
 
         raise ValueError(
